+++ clients/client-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 0.1.18 (2022-07-14)
+
+- Updated search method for a channel with parameter `hidden` and `visbilityList`
+- Raised version of shared modules to 0.1.14
+- Extended get method in base.ts to send data in body e.g. `visbilityList` when searching for channels.
+
 ### 0.1.17 (2022-07-05)
 
 - Updated dependencies